+++ calendar-service/src/main/java/com/example/calendarservice/service/ScheduleService.java
@@ -7,11 +7,11 @@ import java.util.List;
 
 public interface ScheduleService {
     Schedule createSchedule(Schedule schedule);
-    Schedule updateSchedule(Integer scheduleId, Schedule schedule);
-    void deleteSchedule(Integer scheduleId);
-    List<Schedule> getSchedulesByBranchId(Integer branchId, Integer month);
-    List<Schedule> getSchedulesByBranchIdAndUserId(Integer branchId, Integer month, Integer userId);
-    boolean isUserSchedule(Integer scheduleId, int userId);
-    boolean isScheduleInBranch(Integer scheduleId, Integer branchId);
+    Schedule updateSchedule(Long scheduleId, Schedule schedule);
+    void deleteSchedule(Long scheduleId);
+    List<Schedule> getSchedulesByBranchId(Long branchId, Integer month);
+    List<Schedule> getSchedulesByBranchIdAndUserId(Long branchId, Integer month, Long userId);
+    boolean isUserSchedule(Long scheduleId, Long userId);
+    boolean isScheduleInBranch(Long scheduleId, Long branchId);
     List<Schedule> createRepeatSchedules(RepeatScheduleRequest repeatRequest);
 }
